Add javascript experiment
