+++ webapp/src/App.tsx
@@ -28,6 +28,7 @@ import Overview from './components/overview/overview';
 import Graph from './components/dashboard/graph';
 import Calculus from './components/experiments/calculus';
 import Thanks from './components/experiments/thanks';
+import Javascript from './components/experiments/javascript';
 
 function App() {
     return (
@@ -64,6 +65,7 @@ function App() {
             <Route path="experiments">
                 <Route path="thanks" element={<Thanks/>}/>
                 <Route path="calculus" element={<Calculus/>}/>
+                <Route path="javascript" element={<Javascript/>}/>
             </Route>
             <Route element={<RequireAuth allowedRoles={[ROLES.Admin]}/>}>
                 <Route path="/admin" element={<AdminPage/>}/>
