+++ webapp/src/components/experiments/javascript.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import OnePagerWithPicture from './one_pager_with_pic';
+
+const Javascript = () => {
+    return (
+        <OnePagerWithPicture
+            name='Learn Javascript With AI Tutor'
+            description='Master Javascript faster and smarter by learning 5-minute topics daily with our AI tutor'
+            imageURL='/experiments/javascript.jpeg'
+        />
+    );
+}
+
+export default Javascript;
